requirements.txt
selenium>=4.0.0
webdriver-manager>=4.0.0
python-dotenv>=1.0.0
pdfplumber>=0.11.0
google-api-python-client>=2.0.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
